small improvements

+++ include/sys/ringbuf.h
@@ -23,8 +23,8 @@ static inline bool ringbuf_full(ringbuf_t *buf) {
 }
 
 void ringbuf_init(ringbuf_t *rb, void *buf, size_t size);
-void ringbuf_produce(ringbuf_t *buf, unsigned bytes);
-void ringbuf_consume(ringbuf_t *buf, unsigned bytes);
+void ringbuf_produce(ringbuf_t *buf, size_t bytes);
+void ringbuf_consume(ringbuf_t *buf, size_t bytes);
 bool ringbuf_putb(ringbuf_t *buf, uint8_t byte);
 /*! \brief Put exactly n bytes into buf if there's enough space. */
 bool ringbuf_putnb(ringbuf_t *buf, uint8_t *data, size_t n);
