+++ sys/kern/ringbuf.c
@@ -11,8 +11,8 @@ void ringbuf_init(ringbuf_t *rb, void *buf, size_t size) {
 }
 
 /* below two functions are to help to track ringbuf pointers in case when it is
- * changed by something else. e.q. DMA */
-void ringbuf_produce(ringbuf_t *buf, unsigned bytes) {
+ * changed by something else. e.g. DMA */
+void ringbuf_produce(ringbuf_t *buf, size_t bytes) {
   assert(bytes <= buf->size - buf->count);
   buf->count += bytes;
   buf->head += bytes;
@@ -21,7 +21,7 @@ void ringbuf_produce(ringbuf_t *buf, unsigned bytes) {
   assert(buf->head < buf->size);
 }
 
-void ringbuf_consume(ringbuf_t *buf, unsigned bytes) {
+void ringbuf_consume(ringbuf_t *buf, size_t bytes) {
   assert(buf->count >= bytes);
   buf->count -= bytes;
   buf->tail += bytes;
